+++ routes/movies.js
@@ -16,6 +16,10 @@ router.get('/', async (req, res) =>{
 
 //Get ID
 router.get('/:id', async (req, res) =>{
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id))
+        return res.status(404).send('invalid ID');
+        
     const movie = await Movie.findById(req.params.id);
     if (!movie){
         res.status(400)
@@ -57,6 +61,10 @@ router.post('/', auth, async (req, res) =>{
 
 //PUT
 router.put('/:id', auth, async (req, res) =>{
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id))
+        return res.status(404).send('invalid ID');
+        
     const { error } = validate(req.body);
 
     if (error) {
@@ -89,6 +97,10 @@ router.put('/:id', auth, async (req, res) =>{
 
 //Delete
 router.delete('/:id', [auth, admin], async (req, res) =>{
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id))
+        return res.status(404).send('invalid ID');
+        
     const movie = await Movie.findByIdAndRemove(req.params.id);
     if (!movie){
         res.status(400)
